ufshcd: Support HW KM wrapped key mode for fscrypt v2.
porting from https://gerrit.mot.com/#/c/2799638/
rework change to DDK style

When the keymint is supported with wrapped key mode functionalities the UFS
storage driver will retrieve the encrypted keys and send to the secure side
for the underlying provisioning. So is the same with the raw secret for the
inode encryption, the raw key will never be exposed to the HLOS memory but
only the wrapped data will be seen.

Change-Id: I8a500923b0c5ccc9534ae05b771e4ead702b5dc9
Signed-off-by: ouyang feifan <ouyangff@motorola.com>
Reviewed-on: https://gerrit.mot.com/3539577
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Obi-Wan Kenobi
Reviewed-by: Binsheng Que <quebs2@motorola.com>
Submit-Approved: Jira Key